GRHB-35: * rename css files
