+++ frontend/src/components/common/button/button.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'common/types/types';
-import styles from './button.module.scss';
+import styles from './styles.module.scss';
 
 type Props = {
   label: string;
