+++ frontend/src/components/common/link/link.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'common/types/types';
 import { NavLink } from 'react-router-dom';
 import { AppRoute } from 'common/enums/enums';
-import styles from './link.module.scss';
+import styles from './styles.module.scss';
 
 type Props = {
   to: AppRoute;
